streamlit-version/requirements.txt
streamlit==1.30.0
tensorflow==2.15.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
matplotlib==3.7.3